create todoComponent as main one for todoItem, todoForm and Calendar
